Zrównoleglenie - graph generator

+++ FloydWarshall/floyd_warshall_algorithm.h
@@ -1,6 +1,6 @@
 #ifndef _FLOYD_WARSHALL_ALGORITHM_
 #define _FLOYD_WARSHALL_ALGORITHM_
 
-void floyd_warshall(double **graph, int graph_size);
+void floyd_warshall(double **graph, long graph_size);
 
 #endif _FLOYD_WARSHALL_ALGORITHM_
